# Faster merging of hyper c arrays.